8. Стороны прямоугольника. [#484880]

+++ src/ru/job4j/condition/SqArea.java
@@ -3,7 +3,7 @@ package ru.job4j.condition;
 public class SqArea {
     // add method for calculate square
     // p is perimeter, k is coefficient (length is greater then height)
-    public static double square(int p, int k) {
+    public static double square(double p, double k) {
         return Math.pow(p / (2 * (k + 1)), 2) * k;
     }
     // start program
